Add output pin definition to hbridge fan example configuration
To better understand how to use the hbridge fan component, the output pin definitions should be added to the example configuration to make it complete. Otherwise if the user just copy and pasted the example and fills in the pins directly (D1, D2, etc.), he will get an error and not directly notice what he is missing.